Consolidate config option components into single component and file
